--- Nameable.rb
@@ -1,7 +1,0 @@
-class Nameable
-    def correct_name(name)
-        raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
-    end
-    return Nameable
-end
-
