--- base_decorator.rb
@@ -1,11 +1,0 @@
-require_relative 'Nameable'
-class BaseDecorator < Nameable
-    def initialize(nameable_obj)
-        super()
-        @nameable_obj = nameable_obj
-    end
-
-    def correct_name
-        @nameable_obj.correct_name
-    end
-end
