+++ name_able.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# adding nameable class
+class Nameable
+  def correct_name(name)
+    raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
+  end
+end
