+++ output.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require_relative 'person'
+require_relative 'capitalize_decolator'
+require_relative 'trim_decolator'
+
+person = Person.new(22, 'maximilianus')
+p person.correct_name
+capitalized_person = CapitalizeDecorator.new(person)
+p capitalized_person.correct_name
+capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
+p capitalized_trimmed_person.correct_name
